regist and login style + postpage

+++ src/components/Comment.tsx
@@ -3,6 +3,7 @@ import { db } from "../firbase";
 import { Comment as CommentType } from "../constants/types";
 import { useAuth } from "../providers/AuthProvider";
 import { useInfo } from "../providers/InfoProvider";
+import { FaRegTrashAlt } from "react-icons/fa";
 
 type CoP = {
     comment: CommentType
@@ -16,27 +17,30 @@ const Comment = (props: CoP) => {
 
 
     const deleteComment = async (id: string) => {
-        try {
-            const commentRef = doc(collection(db, "comments"), id)
-            await deleteDoc(commentRef);
-        } catch (error) {
-            console.log(error)
+        if (confirm("Delete This Comment ?")) {
+            try {
+                const commentRef = doc(collection(db, "comments"), id)
+                await deleteDoc(commentRef);
+            } catch (error) {
+                console.log(error)
+            }
         }
     }
 
     return (
-        <div key={props.comment.id}>
-            <div className="flex items-center gap-3 p-2">
-                <img className="rounded-full aspect-square" src={commenter?.avatar} width={25} alt="" />
-                <p>{commenter?.username}</p>
+        <div key={props.comment.id} className="bg-gray-100 flex items-center justify-between rounded-md shadow-sm my-4 p-4">
+            <div>
+                <div className="flex items-center gap-3">
+                    <img className="rounded-full w-6 h-6" src={commenter?.avatar} alt="" />
+                    <p className="text-sm">{commenter?.username}</p>
+                </div>
+                <p className="ml-8 mt-1">{props.comment.comment}</p>
             </div>
-            <p className="ml-[35px]">{props.comment.comment}</p>
-            {
-                props.comment.owner == currentUser?.uid ?
-                    <button onClick={() => { deleteComment(props.comment.id) }}>Delete Comment</button>
-                    :
-                    null
-            }
+            {props.comment.owner === currentUser?.uid && (
+                <button className="text-red-600 text-sm mt-1" onClick={() => { deleteComment(props.comment.id) }}>
+                    <FaRegTrashAlt />
+                </button>
+            )}
         </div>
     )
 }
